repr_offset
Module for_examples

Types used for examples,

These are in the docs purely so that documentation examples only use types that are documented.

You can only use items from this module when the “for_examples” feature is enabled.

Structs

ReprAlign4

An example #[repr(C, align(4))] type

ReprC

An example #[repr(C)] type

ReprCNoGFO

An example #[repr(C)] type which doesn’t implement GetFieldOffset

ReprPacked

An example #[repr(C, packed)] type

ReprPacked2

An example #[repr(C, packed(2))] type.