repr_offset::get_field_offset

Struct FieldOffsetWithVis

pub struct FieldOffsetWithVis<S, V, FN, F, A> { /* private fields */ }

A wrapper around a FieldOffset, with a visibility type parameter (whether the field is pub or not).

Type parameters

S: The type that contains the field

V: The visibility of the field, either IsPrivate or IsPublic.

FN: The name of the field, written using the repr_offset::tstr::TS macro, written as TS!(field_name).

F: The type of the field.

A: The alignment of the field inside of S, either Aligned or Unaligned.

Implementations

impl<S, V, FN, F, A> FieldOffsetWithVis<S, V, FN, F, A>

pub const unsafe fn new(offset: usize) -> Self

Constructs this FieldOffsetWithVis from the offset (in bytes) of a field.

Safety

This method has a superset of the safety requirements of FieldOffset::new.

The V type parameter must be:

IsPublic: When the field is pub.

IsPrivate: When the field has the default (private) visibility, or has a visibility smaller or equal to pub(crate).

The FN type parameter must be the name of the field using the repr_offset::tstr::TS macro, eg: TS!(foo) for the foo field.

pub const unsafe fn from_fieldoffset(offset: FieldOffset<S, F, A>) -> Self

Constructs this FieldOffsetWithVis from offset.

Safety

The V type parameter must be:

IsPublic: When the field is pub.

IsPrivate: When the field has the default (private) visibility, or has a visibility smaller or equal to pub(crate).

The FN type parameter must be the name of the field using the repr_offset::tstr::TS macro, eg: TS!(foo) for the foo field.

impl<FN, S, F, A> FieldOffsetWithVis<S, IsPublic, FN, F, A>

pub const fn to_field_offset(self) -> FieldOffset<S, F, A>

Unwraps this into a FieldOffset for a public field.

impl<S, V, FN, F, A> FieldOffsetWithVis<S, V, FN, F, A>

pub const unsafe fn private_field_offset(self) -> FieldOffset<S, F, A>

Unwraps this into a FieldOffset for a possibly private field.

Safety

Because the field may be private, modifying its state could cause undefined behavior, and is only supposed to be done in a context where the field is accessible.

pub const unsafe fn cast_struct<SO>(self) -> FieldOffsetWithVis<SO, V, FN, F, A>

Casts this FieldOffsetWithVis to be for a different struct.

This is mostly useful for #[repr(transparent)] types to delegate to their single field.

Safety

SO must contain a field of type S as its first field in memory, at offset 0.