Trait repr_offset::get_field_offset::ImplsGetFieldOffset

pub unsafe trait ImplsGetFieldOffset: Sized { }

Marker trait for types that implement GetFieldOffset.

This trait is required for the GetFieldOffset impls that get the FieldOffset of nested fields.

This is only required as a workaround to lower the time that cargo doc takes to run.