Module repr_offset::alignment

Type-level encoding of enum Alignment { Aligned, Unaligned }

Structs

Aligned

A marker type representing that a FieldOffset is for an aligned field.

Unaligned

A marker type representing that a FieldOffset is for a (potentially) unaligned field.

Traits

Alignment

Marker trait for types that represents the alignment of a FieldOffset.

CombineAlignment

Trait that combines two Alignment types, determines the return type of FieldOffset + FieldOffset.

Type Aliases

CombineAlignmentOut

Combines two Alignment types, determines the return type of FieldOffset + FieldOffset.