repr_offset::alignment

Trait CombineAlignment

pub trait CombineAlignment<Rhs: Alignment> {
    type Output: Alignment;
}

Trait that combines two Alignment types, determines the return type of FieldOffset + FieldOffset.

Required Associated Types

type Output: Alignment

This is Aligned if both Self and the Rhs parameter are Aligned, otherwise it is Unaligned.