repr_offset::get_field_offset

Trait ImplsGetFieldOffset

pub unsafe trait ImplsGetFieldOffset: Sized { }

Marker trait for types that implement GetFieldOffset.

This trait is required for the GetFieldOffset impls that get the FieldOffset of nested fields.

This is only required as a workaround to lower the time that cargo doc takes to run.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.